_Jeeves_1.0/requirements.txt
google-auth
google-auth-oauthlib==1.0.0
google-auth-httplib2
google-api-python-client